削除モーダル表示

+++ app/Http/Controllers/PostsController.php
@@ -69,10 +69,16 @@ class PostsController extends Controller
     //投稿の削除
     public function postDelete($id){
 
-        dd($id);
-        $post = Post::where('id', $id);
-        $post->delete();
-        return redirect('/top');
+        //IDに基づいて投稿を取得
+        //dd($id);
+        $post = Post::find($id);
+
+        // 投稿が見つかった場合のみ削除処理を行う
+        if ($post) {
+            $post->delete();
+        }
+
+        return redirect()->route('top');
 
     }
